fix(control-ui): split document preview module and e2e file under lint caps
